+++ modules/agent/src/main/java/io/jpom/socket/AgentWebSocketConfig.java
@@ -22,9 +22,21 @@
  */
 package io.jpom.socket;
 
-import org.springframework.context.annotation.Bean;
+import org.springframework.beans.factory.InitializingBean;
+import org.springframework.beans.factory.SmartInitializingSingleton;
+import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.socket.server.standard.ServerEndpointExporter;
+import org.springframework.util.Assert;
+import org.springframework.web.context.support.WebApplicationObjectSupport;
+
+import javax.websocket.DeploymentException;
+import javax.websocket.server.ServerContainer;
+import javax.websocket.server.ServerEndpoint;
+import javax.websocket.server.ServerEndpointConfig;
+import java.util.LinkedHashSet;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
 
 /**
  * 插件端socket 配置
@@ -33,10 +45,91 @@ import org.springframework.web.socket.server.standard.ServerEndpointExporter;
  * @since 2019/4/19
  */
 @Configuration
-public class AgentWebSocketConfig {
+public class AgentWebSocketConfig extends WebApplicationObjectSupport implements InitializingBean, SmartInitializingSingleton {
 
-	@Bean
-	public ServerEndpointExporter serverEndpointExporter() {
-		return new ServerEndpointExporter();
-	}
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        //Assert.state(this.serverContainer != null, "javax.websocket.server.ServerContainer not available");
+    }
+
+
+//    @Override
+//    protected void initServletContext(ServletContext servletContext) {
+//        if (this.serverContainer == null) {
+//            this.serverContainer = (ServerContainer) servletContext.getAttribute("javax.websocket.server.ServerContainer");
+//        }
+//    }
+
+    private ServerContainer getServerContainer() {
+        return (ServerContainer) Objects.requireNonNull(getServletContext()).getAttribute("javax.websocket.server.ServerContainer");
+    }
+
+    @Override
+    protected boolean isContextRequired() {
+        return true;
+    }
+
+    @Override
+    public void afterSingletonsInstantiated() {
+        registerEndpoints();
+    }
+
+    protected void registerEndpoints() {
+        Set<Class<?>> endpointClasses = new LinkedHashSet<>();
+
+        ApplicationContext context = getApplicationContext();
+        if (context != null) {
+            String[] endpointBeanNames = context.getBeanNamesForAnnotation(ServerEndpoint.class);
+            for (String beanName : endpointBeanNames) {
+                endpointClasses.add(context.getType(beanName));
+            }
+        }
+
+        for (Class<?> endpointClass : endpointClasses) {
+            registerEndpoint(endpointClass);
+        }
+
+        if (context != null) {
+            Map<String, ServerEndpointConfig> endpointConfigMap = context.getBeansOfType(ServerEndpointConfig.class);
+            for (ServerEndpointConfig endpointConfig : endpointConfigMap.values()) {
+                registerEndpoint(endpointConfig);
+            }
+        }
+    }
+
+    private void registerEndpoint(Class<?> endpointClass) {
+        ServerContainer serverContainer = getServerContainer();
+        Assert.state(serverContainer != null,
+            "No ServerContainer set. Most likely the server's own WebSocket ServletContainerInitializer " +
+                "has not run yet. Was the Spring ApplicationContext refreshed through a " +
+                "org.springframework.web.context.ContextLoaderListener, " +
+                "i.e. after the ServletContext has been fully initialized?");
+        try {
+            if (logger.isDebugEnabled()) {
+                logger.debug("Registering @ServerEndpoint class: " + endpointClass);
+            }
+            serverContainer.addEndpoint(endpointClass);
+        } catch (DeploymentException ex) {
+            throw new IllegalStateException("Failed to register @ServerEndpoint class: " + endpointClass, ex);
+        }
+    }
+
+    private void registerEndpoint(ServerEndpointConfig endpointConfig) {
+        ServerContainer serverContainer = this.getServerContainer();
+        Assert.state(serverContainer != null, "No ServerContainer set");
+        try {
+            if (logger.isDebugEnabled()) {
+                logger.debug("Registering ServerEndpointConfig: " + endpointConfig);
+            }
+            serverContainer.addEndpoint(endpointConfig);
+        } catch (DeploymentException ex) {
+            throw new IllegalStateException("Failed to register ServerEndpointConfig: " + endpointConfig, ex);
+        }
+    }
+
+//	@Bean
+//	public ServerEndpointExporter serverEndpointExporter() {
+//		return new ServerEndpointExporter();
+//	}
 }