+++ CHANGELOG.md
@@ -17,6 +17,7 @@
    （感谢 [@LeonChen21](https://gitee.com/leonchen21) [Gitee issues I67C3C](https://gitee.com/dromara/Jpom/issues/I67C3C) ）
 7. 【server】优化 websocket 控制台操作日志记录
 8. 【server】修复 超级管理的 websocket 操作日志记录工作空间不正确
+9. 【agent】优化 插件端删除 spring-boot-starter-websocket 依赖
 
 ### ❌ 不兼容功能